+++ x/erc20/keeper/msg_server.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"bytes"
 	"context"
 	"math/big"
 
@@ -33,7 +34,7 @@ func (k Keeper) ConvertERC20(
 	receiver := sdk.MustAccAddressFromBech32(msg.Receiver)
 	sender := common.HexToAddress(msg.Sender)
 
-	pair, err := k.MintingEnabled(ctx, sender.Bytes(), receiver, msg.ContractAddress)
+	pair, err := k.MintingEnabled(ctx, receiver, msg.ContractAddress, bytes.Equal(sender.Bytes(), receiver))
 	if err != nil {
 		return nil, err
 	}
@@ -199,7 +200,7 @@ func (k Keeper) ConvertCoin(
 	sender := sdk.MustAccAddressFromBech32(msg.Sender)
 	receiver := common.HexToAddress(msg.Receiver)
 
-	pair, err := k.MintingEnabled(ctx, sender, receiver.Bytes(), msg.Coin.Denom)
+	pair, err := k.MintingEnabled(ctx, receiver.Bytes(), msg.Coin.Denom, bytes.Equal(sender, receiver.Bytes()))
 	if err != nil {
 		return nil, err
 	}