+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### BUG FIXES
 
 - [\#748](https://github.com/cosmos/evm/pull/748) Fix DynamicFeeChecker in Cosmos ante handler to respect NoBaseFee feemarkets' parameter.
+- [\#]() Fix erc20 ibc middleware to not to validate sender address format.
 
 ## v0.5.0